requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic-settings
passlib[bcrypt]
PyJWT
python-multipart
httpx
python-dotenv
pymupdf
requests
selenium
webdriver-manager
apscheduler
sqlmodel
email-validator
playwright>=1.44,<2.0
beautifulsoup4>=4.12,<5.0
lxml>=5,<6